requirements.txt
streamlit
langchain-community
sentence-transformers
faiss-cpu
transformers
torch
pypdf